requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
google-auth
google-auth-oauthlib
google-api-python-client
pdfplumber
python-docx
python-multipart
jinja2
itsdangerous
fpdf2
